Refactor room sharing functions into separate modules for better maintainability
